+++ src/Builder/FallbackBuilder.php
@@ -28,14 +28,14 @@ use Ramsey\Uuid\UuidInterface;
 class FallbackBuilder implements UuidBuilderInterface
 {
     /**
-     * @var UuidBuilderInterface[]
+     * @var BuilderCollection
      */
-    private $builders = [];
+    private $builders;
 
     /**
-     * @param UuidBuilderInterface[] $builders An array of UUID builders
+     * @param BuilderCollection $builders An array of UUID builders
      */
-    public function __construct(array $builders)
+    public function __construct(BuilderCollection $builders)
     {
         $this->builders = $builders;
     }
@@ -53,6 +53,7 @@ class FallbackBuilder implements UuidBuilderInterface
     {
         $lastBuilderException = null;
 
+        /** @var UuidBuilderInterface $builder */
         foreach ($this->builders as $builder) {
             try {
                 return $builder->build($codec, $bytes);